Changed: Active downloads page has been moved away from the dashboard
