+++ Mnema.Frontend/Web/public/assets/i18n/en.json
@@ -22,6 +22,7 @@
     },
     "actions": {
       "title": "Actions",
+      "downloads": "Downloads",
       "subscriptions": "Subscriptions",
       "notifications": "Notifications",
       "audit-log": "Audit log"
