+++ Mnema.Frontend/Web/src/app/_routes/extra.routes.ts
@@ -1,6 +1,7 @@
 import {Routes} from "@angular/router";
 import {SubscriptionManagerComponent} from "../subscription-manager/subscription-manager.component";
 import {NotificationsComponent} from "../notifications/notifications.component";
+import {ActiveDownloadsComponent} from "../dashboard/active-downloads/active-downloads.component";
 
 export const routes: Routes = [
   {
@@ -10,5 +11,9 @@ export const routes: Routes = [
   {
     path: 'notifications',
     component: NotificationsComponent
+  },
+  {
+    path: 'active-downloads',
+    component: ActiveDownloadsComponent
   }
 ]
