+++ Mnema.Frontend/Web/src/app/dashboard/dashboard.component.html
@@ -1,175 +1,3 @@
-@if (this.loading()) {
-  <app-loading-spinner [centered]="true" size="large"></app-loading-spinner>
-} @else {
-  <div class="w-100" *transloco="let t;prefix:'dashboard'">
-    @if (dashboardItems().length === 0) {
-      <app-button-grid [groups]="buttonGroupService.dashboardGroups()" />
-    } @else {
-      <div class="pt-1 pt-md-3 m-3">
-
-        <div class="d-flex flex-grow-1 justify-content-end mb-2">
-          <button class="btn btn-secondary" (click)="manualAdd()">
-            {{t('manual-add.title')}}
-          </button>
-        </div>
-
-        <app-table
-          [items]="dashboardItems()"
-          [trackByIdFunc]="itemTrackBy"
-          [noHoverColour]="true"
-          *transloco="let t;prefix:'dashboard.table'"
-        >
-          <ng-template #mobile let-info>
-            <div class="card mobile-card">
-              <div class="card-header">
-                @if (info.refUrl === "") {
-                  <h2 class="card-title h5 mb-0">{{ info.name | contentTitle }}</h2>
-                } @else {
-                  <a class="card-title h5 mb-0" target="_blank" rel="noopener noreferrer"
-                     [href]="info.refUrl">{{ info.name | contentTitle }}</a>
-                }
-              </div>
-
-              <div class="card-body">
-                <div class="card-row">
-                  <span class="card-label">{{ t('size') }}</span>
-                  <span class="card-value fw-semibold">{{ info.size }}</span>
-                </div>
-
-                <div class="card-row">
-                  <span class="card-label">{{ t('status') }}</span>
-                  <span class="card-value">
-                    <app-badge colour="primary">{{info.contentState | contentState}}</app-badge>
-                  </span>
-                </div>
-
-                <div class="card-progress">
-                  <div class="progress" style="height: 8px;">
-                    <div
-                      class="progress-bar"
-                      role="progressbar"
-                      [style.background-color]="'var(--secondary-color)'"
-                      [style.width.%]="info.progress"
-                      [attr.aria-valuenow]="info.progress"
-                      aria-valuemin="0"
-                      aria-valuemax="100"
-                    ></div>
-                  </div>
-                  <div class="card-progress-info">
-                    <span class="card-progress-text">{{ info.speed | speed:info.speedType }} {{ info.speedType | speedType }}</span>
-                    @if (info.estimated) {
-                      <span class="card-progress-text">{{ info.estimated | time }}</span>
-                    }
-                    <span class="card-progress-text">{{ info.progress }}%</span>
-                  </div>
-                </div>
-              </div>
-
-              <div class="card-footer">
-                <div class="card-actions">
-                  @if (info.contentState == ContentState.Downloading) {
-                    <button (click)="browse(info)" class="btn btn-primary" [ngbTooltip]="t('actions.browse')">
-                      <span class="fa fa-folder"></span>
-                    </button>
-                  }
-
-                  @if (info.contentState == ContentState.Waiting) {
-                    <button (click)="pickContent(info)" class="btn btn-primary" [ngbTooltip]="t('actions.set-content')" >
-                      <span class="fa fa-cart-plus"></span>
-                    </button>
-
-                    <button (click)="markReady(info)" class="btn btn-secondary" [ngbTooltip]="t('actions.mark-ready')" >
-                      <span class="fa fa-check"></span>
-                    </button>
-                  }
-
-                  @if (info.contentState !== ContentState.Cleanup) {
-                    <button (click)="stop(info)" class="btn btn-error" [ngbTooltip]="t('actions.stop')" >
-                      <span class="fa fa-trash"></span>
-                    </button>
-                  }
-                </div>
-              </div>
-            </div>
-          </ng-template>
-
-          <ng-template #header>
-            <tr>
-              <th style="width:20%" class="table-header-cell">{{ t('name') }}</th>
-              <th style="width:10%" class="table-header-cell">{{ t('size') }}</th>
-              <th style="width:30%" class="table-header-cell">{{ t('progress') }}</th>
-              <th style="width:10%" class="table-header-cell">{{ t('status') }}</th>
-              <th style="width:15%" class="table-header-cell">{{ t('actions.label') }}</th>
-            </tr>
-          </ng-template>
-
-          <ng-template #cell let-info>
-            <th class="table-cell">
-              @if (info.refUrl === "") {
-                <h2 class="h5">{{ info.name | contentTitle }}</h2>
-              } @else {
-                <a class="h5 text-decoration-none" target="_blank" rel="noopener noreferrer"
-                   [href]="info.refUrl">{{ info.name | contentTitle }}</a>
-              }
-            </th>
-            <th class="table-cell">
-              <span class="fw-semibold">{{ info.size }}</span>
-            </th>
-            <th class="table-cell">
-              <div class="position-relative pt-1 m-2">
-                <div class="progress" style="height: 8px;">
-                  <div
-                    class="progress-bar"
-                    role="progressbar"
-                    [style.background-color]="'var(--secondary-color)'"
-                    [style.width.%]="info.progress"
-                    [attr.aria-valuenow]="info.progress"
-                    aria-valuemin="0"
-                    aria-valuemax="100"
-                  ></div>
-                </div>
-                <div class="d-flex flex-column flex-md-row justify-content-between small text-muted mt-1">
-                  <span>{{ info.speed | speed:info.speedType }} {{ info.speedType | speedType }}</span>
-                  @if (info.estimated) {
-                    <span class="d-none d-md-inline">{{ info.estimated | time }}</span>
-                  }
-                  <span>{{ info.progress }}%</span>
-                </div>
-              </div>
-            </th>
-            <th class="table-cell">
-              <app-badge colour="primary">{{info.contentState | contentState}}</app-badge>
-            </th>
-            <th class="table-cell">
-              <div class="d-flex flex-column flex-md-row gap-2 my-2">
-
-                @if (info.contentState == ContentState.Downloading) {
-                  <button (click)="browse(info)" class="btn btn-primary" [ngbTooltip]="t('actions.browse')">
-                    <span class="fa fa-folder"></span>
-                  </button>
-                }
-
-                @if (info.contentState == ContentState.Waiting) {
-                  <button (click)="pickContent(info)" class="btn btn-primary" [ngbTooltip]="t('actions.set-content')" >
-                    <span class="fa fa-cart-plus"></span>
-                  </button>
-
-                  <button (click)="markReady(info)" class="btn btn-secondary" [ngbTooltip]="t('actions.mark-ready')" >
-                    <span class="fa fa-check"></span>
-                  </button>
-                }
-
-                @if (info.contentState !== ContentState.Cleanup) {
-                  <button (click)="stop(info)" class="btn btn-error" [ngbTooltip]="t('actions.stop')" >
-                    <span class="fa fa-trash"></span>
-                  </button>
-                }
-
-              </div>
-            </th>
-          </ng-template>
-        </app-table>
-      </div>
-    }
-  </div>
-}
+<div class="w-100" *transloco="let t;prefix:'dashboard'">
+  <app-button-grid [groups]="buttonGroupService.dashboardGroups()" />
+</div>
